use consistent authentication flow vs authorization flow terminology
The IndieAuth specification describes two flows: the Authentication
flow, and the Authorization flow.

This change updates a few more places where the wrong terms were used,
improving consistency.

The term "Authorization Code Flow" is also used, but it is an OAuth 2.0
term. Make it more clear by prepending "OAuth 2.0" to the phrase.

Updates #13