use consistent authentication request vs authorization request terminology
In the context of an authentication flow, there is an authentication
request and response. In the context of an authorization flow, there
is an authorization request and response. Fix up a few more places
where wrong term was used.

A few places that refer to both flows were mentioning the authorization
request explicitly. Update that text to make it clear it applies to both
authentication and authorization requests.

Also be more consistent about the order. Always place authentication first
and authorization second, whenever both are mentioned.

Updates #13